requirements.txt
djangorestframework~=3.12.4
PyJWT~=1.7.0
Django~=3.2.3
djangorestframework-jwt~=1.11.0